Find most probable motif sequences in a proteome
Edit motif profile (lines 1-45) and/or name of proteome file (line 162).

Commented out section between 170 and 172 prints the top 25 scoring
proteins.
Line 174 generates the boxplot.
The boxplot function can take multiple proteomes as an array of
dictionary values, and the boxplot labels need to be specified as an
array of strings as the second argument of boxplot.